chore: update configuration and database initialization scripts
Updated the .air.toml configuration file to modify the exclusion
regex key and adjust the kill delay to 60 seconds. Enabled sending
interrupts and stopping on errors to improve process handling.

In the .docker/init.sh script, added the creation of the "pg_cron"
extension for the sesamo_test database and included the "hstore"
extension for all databases. These changes enhance database
functionality and ensure necessary extensions are available.