feat: update Dockerfile to support PostgreSQL 16 and add pgx_ulid
The Dockerfile for PostgreSQL has been updated to support version 16,
including the addition of the pgx_ulid extension. The previous Alpine
Linux base image has been replaced with a Debian-based image to
facilitate the installation of the pgx_ulid package. The pg_cron
extension is now installed using the package manager, and the
Dockerfile now includes commands to download and install the pgx_ulid
extension. This change enhances the database capabilities by allowing
the use of ULIDs and improves compatibility with the latest PostgreSQL
version.